added home and fullscreen button rendering changes, toggling between fullscreen and quadrant views still works only once
